feat(menu): split out menubar functionality into it's own class
Splitting out Menubar into it's own class now gives the option for extending more menu types

BREAKING CHANGE: AccessibleMenu is no longer a constructor, so uses will have to use
AccessibleMenu.Menubar

See #61
