+++ src/_baseMenu.js
@@ -1,16 +1,16 @@
-import MenuItem from "./menuItem";
 import MenuToggle from "./menuToggle";
-import { keyPress, preventEvent } from "./eventHandlers";
+import MenuItem from "./menuItem";
 import {
   isHTMLElement,
   isCSSSelector,
-  isBoolean,
-  isNumber,
   isString,
+  isBoolean,
   isMenu,
+  isNumber,
   isValidState,
   isValidEvent,
 } from "./validate";
+import { preventEvent } from "./eventHandlers";
 
 /**
  * An accessible navigation element in the DOM.
@@ -110,7 +110,7 @@ class BaseMenu {
   }
 
   /**
-   * Initializes the menu with proper tab indexing and properties.
+   * Initializes the menu.
    *
    * This will also initialize all menu items and sub menus.
    */
@@ -120,13 +120,8 @@ class BaseMenu {
 
     // Set all of the DOM elements.
     this.setDOMElements();
-    this.dom.menu.setAttribute("role", "menubar");
-    this.createChildElements();
 
     if (this.isTopLevel) {
-      // Set initial tabIndex.
-      this.currentMenuItem.dom.link.tabIndex = 0;
-
       if (this.dom.controller && this.dom.container) {
         // Create a new MenuToggle to control the menu.
         const toggle = new MenuToggle({
@@ -140,12 +135,6 @@ class BaseMenu {
         this.menuElements.controller = toggle;
       }
     }
-
-    this.handleFocus();
-    this.handleClick();
-    if (this.isHoverable) this.handleHover();
-    this.handleKeydown();
-    this.handleKeyup();
   }
 
   /**
@@ -449,7 +438,7 @@ class BaseMenu {
   }
 
   /**
-   * Finds the root Menu element.
+   * Finds the root menu element.
    *
    * @param {BaseMenu} menu - The menu to check.
    */
@@ -465,8 +454,10 @@ class BaseMenu {
 
   /**
    * Creates and initializes all menu items and submenus.
+   *
+   * @param {object} MenuType - The menu type for created submenus.
    */
-  createChildElements() {
+  createChildElements(MenuType = BaseMenu) {
     this.dom.menuItems.forEach(element => {
       let menuItem;
 
@@ -476,8 +467,8 @@ class BaseMenu {
         // The actual menu DOM element.
         const submenu = element.querySelector(this.selectors.submenus);
 
-        // Create the new Menu and initialize it.
-        const menu = new BaseMenu({
+        // Create the new menu and initialize it.
+        const menu = new MenuType({
           menuElement: submenu,
           menuItemSelector: this.selectors.menuItems,
           menuLinkSelector: this.selectors.menuLinks,
@@ -547,291 +538,10 @@ class BaseMenu {
   }
 
   /**
-   * Handles keydown events throughout the menu for proper menu use.
-   */
-  handleKeydown() {
-    this.dom.menu.addEventListener("keydown", event => {
-      this.currentEvent = "keyboard";
-
-      const key = keyPress(event);
-
-      if (key === "Tab") {
-        // Hitting Tab:
-        // - Moves focus out of the menu.
-        if (this.elements.rootMenu.focusState !== "none") {
-          this.elements.rootMenu.blur();
-          this.elements.rootMenu.closeChildren();
-        } else {
-          this.elements.rootMenu.focus();
-        }
-      }
-
-      // Prevent default event actions if we're handling the keyup event.
-      if (key === "Character") {
-        preventEvent(event);
-      } else if (this.isTopLevel) {
-        if (this.focusState === "self") {
-          const keys = ["ArrowRight", "ArrowLeft", "Home", "End"];
-          const submenuKeys = ["Space", "Enter", "ArrowDown", "ArrowUp"];
-          const controllerKeys = ["Espace"];
-
-          if (keys.includes(key)) {
-            preventEvent(event);
-          } else if (
-            this.currentMenuItem.isSubmenuItem &&
-            submenuKeys.includes(key)
-          ) {
-            preventEvent(event);
-          } else if (this.elements.controller && controllerKeys.includes(key)) {
-            preventEvent(event);
-          }
-        }
-      } else {
-        const keys = [
-          "Escape",
-          "ArrowRight",
-          "ArrowLeft",
-          "ArrowDown",
-          "ArrowUp",
-          "Home",
-          "End",
-        ];
-        const submenuKeys = ["Space", "Enter"];
-
-        if (keys.includes(key)) {
-          preventEvent(event);
-        } else if (
-          this.currentMenuItem.isSubmenuItem &&
-          submenuKeys.includes(key)
-        ) {
-          preventEvent(event);
-        }
-      }
-    });
-
-    if (this.isTopLevel && this.elements.controller) {
-      this.elements.controller.dom.toggle.addEventListener("keydown", () => {
-        this.currentEvent = "keyboard";
-
-        const key = keyPress(event);
-
-        if (key === "Space" || key === "Enter") {
-          preventEvent(event);
-        }
-      });
-    }
-  }
-
-  /**
-   * Handles keyup events throughout the menu for proper menu use.
-   */
-  handleKeyup() {
-    this.dom.menu.addEventListener("keyup", event => {
-      this.currentEvent = "keyboard";
-
-      const key = keyPress(event);
-      const { altKey, crtlKey, metaKey } = event;
-      const modifier = altKey || crtlKey || metaKey;
-
-      if (key === "Character" && !modifier) {
-        // Hitting Character:
-        // - Moves focus to next item in the menubar having a name that starts with the typed character.
-        // - If none of the items have a name starting with the typed character, focus does not move.
-        preventEvent(event);
-        this.focusNextChildWithCharacter(event.key);
-      } else if (this.isTopLevel) {
-        if (this.focusState === "self") {
-          if (key === "Space" || key === "Enter") {
-            // Hitting Space or Enter:
-            // - Opens submenu and moves focus to first item in the submenu.
-            if (this.currentMenuItem.isSubmenuItem) {
-              preventEvent(event);
-              this.currentMenuItem.elements.toggle.open();
-              this.currentMenuItem.elements.childMenu.focusFirstChild();
-            }
-          } else if (key === "ArrowRight") {
-            // Hitting the Right Arrow:
-            // - Moves focus to the next item in the menubar.
-            // - If focus is on the last item, moves focus to the first item.
-            // - If focus was on an open submenu and the newly focussed item has a submenu, open the submenu.
-            preventEvent(event);
-
-            // Store the current item's info if its an open dropdown.
-            const previousChildOpen =
-              this.currentMenuItem.isSubmenuItem &&
-              this.currentMenuItem.elements.toggle.isOpen;
-
-            this.focusNextChild();
-
-            // Open the newly focussed submenu if applicable.
-            if (previousChildOpen) {
-              if (this.currentMenuItem.isSubmenuItem) {
-                this.currentMenuItem.elements.toggle.preview();
-              } else {
-                this.closeChildren();
-              }
-            }
-          } else if (key === "ArrowLeft") {
-            // Hitting the Left Arrow:
-            // - Moves focus to the previous item in the menubar.
-            // - If focus is on the first item, moves focus to the last item.
-            // - If focus was on an open submenu and the newly focussed item has a submenu, open the submenu.
-            preventEvent(event);
-
-            // Store the current item's info if its an open dropdown.
-            const previousChildOpen =
-              this.currentMenuItem.isSubmenuItem &&
-              this.currentMenuItem.elements.toggle.isOpen;
-
-            this.focusPreviousChild();
-
-            // Open the newly focussed submenu if applicable.
-            if (previousChildOpen) {
-              if (this.currentMenuItem.isSubmenuItem) {
-                this.currentMenuItem.elements.toggle.preview();
-              } else {
-                this.closeChildren();
-              }
-            }
-          } else if (key === "ArrowDown") {
-            // Hitting the Down Arrow:
-            // - Opens submenu and moves focus to first item in the submenu.
-            if (this.currentMenuItem.isSubmenuItem) {
-              preventEvent(event);
-              this.currentMenuItem.elements.toggle.open();
-              this.currentMenuItem.elements.childMenu.focusFirstChild();
-            }
-          } else if (key === "ArrowUp") {
-            // Hitting the Up Arrow:
-            // - Opens submenu and moves focus to last item in the submenu.
-            if (this.currentMenuItem.isSubmenuItem) {
-              preventEvent(event);
-              this.currentMenuItem.elements.toggle.open();
-              this.currentMenuItem.elements.childMenu.focusLastChild();
-            }
-          } else if (key === "Home") {
-            // Hitting Home:
-            // - Moves focus to first item in the menubar.
-            preventEvent(event);
-            this.focusFirstChild();
-          } else if (key === "End") {
-            // Hitting End:
-            // - Moves focus to last item in the menubar.
-            preventEvent(event);
-            this.focusLastChild();
-          } else if (key === "Escape") {
-            if (this.elements.controller !== null) {
-              // Hitting Escape:
-              // - Closes menu.
-              if (this.elements.controller.isOpen) {
-                preventEvent(event);
-                this.elements.controller.close();
-                this.focusController();
-              }
-            }
-          }
-        }
-      } else {
-        if (key === "Space" || key === "Enter") {
-          // Hitting Space or Enter:
-          // - Activates menu item, causing the link to be activated.
-          if (this.currentMenuItem.isSubmenuItem) {
-            preventEvent(event);
-            this.currentMenuItem.elements.toggle.open();
-            this.currentMenuItem.elements.childMenu.focusFirstChild();
-          }
-        } else if (key === "Escape") {
-          // Hitting Escape:
-          // - Closes submenu.
-          // - Moves focus to parent menubar item.
-          preventEvent(event);
-          this.elements.rootMenu.closeChildren();
-          this.elements.rootMenu.focusCurrentChild();
-        } else if (key === "ArrowRight") {
-          // Hitting the Right Arrow:
-          // - If focus is on an item with a submenu, opens the submenu and places focus on the first item.
-          // - If focus is on an item that does not have a submenu:
-          //   - Closes submenu.
-          //   - Moves focus to next item in the menubar.
-          //   - Opens submenu of newly focused menubar item, keeping focus on that parent menubar item.
-          if (this.currentMenuItem.isSubmenuItem) {
-            preventEvent(event);
-            this.currentMenuItem.elements.toggle.open();
-            this.currentMenuItem.elements.childMenu.focusFirstChild();
-          } else {
-            preventEvent(event);
-            this.elements.rootMenu.closeChildren();
-            this.elements.rootMenu.focusNextChild();
-
-            if (this.elements.rootMenu.currentMenuItem.isSubmenuItem) {
-              this.elements.rootMenu.currentMenuItem.elements.toggle.preview();
-            }
-          }
-        } else if (key === "ArrowLeft") {
-          // Hitting the Left Arrow:
-          // - Closes submenu and moves focus to parent menu item.
-          // - If parent menu item is in the menubar, also:
-          //   - moves focus to previous item in the menubar.
-          //   - Opens submenu of newly focused menubar item, keeping focus on that parent menubar item.
-          if (this.elements.parentMenu.currentMenuItem.isSubmenuItem) {
-            preventEvent(event);
-            this.elements.parentMenu.currentMenuItem.elements.toggle.close();
-            this.elements.parentMenu.focusCurrentChild();
-
-            if (this.elements.parentMenu === this.elements.rootMenu) {
-              this.elements.rootMenu.closeChildren();
-              this.elements.rootMenu.focusPreviousChild();
-
-              if (this.elements.rootMenu.currentMenuItem.isSubmenuItem) {
-                this.elements.rootMenu.currentMenuItem.elements.toggle.preview();
-              }
-            }
-          }
-        } else if (key === "ArrowDown") {
-          // Hitting the Down Arrow:
-          // - Moves focus to the next item in the menubar.
-          // - If focus is on the last item, moves focus to the first item.
-          preventEvent(event);
-          this.focusNextChild();
-        } else if (key === "ArrowUp") {
-          // Hitting the Up Arrow:
-          // - Moves focus to the previous item in the menubar.
-          // - If focus is on the first item, moves focus to the last item.
-          preventEvent(event);
-          this.focusPreviousChild();
-        } else if (key === "Home") {
-          // Hitting Home:
-          // - Moves focus to first item in the menubar.
-          preventEvent(event);
-          this.focusFirstChild();
-        } else if (key === "End") {
-          // Hitting End:
-          // - Moves focus to last item in the menubar.
-          preventEvent(event);
-          this.focusLastChild();
-        }
-      }
-    });
-
-    if (this.isTopLevel && this.elements.controller) {
-      this.elements.controller.dom.toggle.addEventListener("keyup", () => {
-        this.currentEvent = "keyboard";
-
-        const key = keyPress(event);
-
-        if (key === "Space" || key === "Enter") {
-          preventEvent(event);
-          this.elements.controller.open();
-          this.focusFirstChild();
-        }
-      });
-    }
-  }
-
-  /**
    * Handles click events throughout the menu for proper use.
    */
   handleClick() {
+    // Close the menu if a click event happens outside of it.
     document.addEventListener("click", event => {
       if (this.focusState !== "none") {
         this.currentEvent = "mouse";
